OS211 machffud

+++ TXT/myscript.sh
@@ -1,29 +1,55 @@
 #!/bin/bash
-#Machffud Tra H V
+# Copyright (C) 2020-2021 Cicak Bin Kadal
+# https://www.youtube.com/watch?v=KAXK07ni9gU
 
-FILES="my*.txt my*.sh"
+# This free document is distributed in the hope that it will be 
+# useful, but WITHOUT ANY WARRANTY; without even the implied 
+# warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+# REV02 Fri 12 Mar 13:40:58 WIB 2021
+# REV01 Tue 13 Oct 10:37:14 WIB 2020
+# START Mon 28 Sep 21:05:04 WIB 2020
+
+REC1="operatingsystems@vlsm.org"
+REC2="machffud.thv@gmail.com"
+FILES="my*.asc my*.txt my*.sh"
 SHA="SHA256SUM"
 
-echo "rm -f $SHA $SHA.asc"
-rm -f $SHA $SHA.asc
+[ -d $HOME/RESULT ] || { echo "No $HOME/RESULT directory" ; exit; }
+pushd $HOME/RESULT
+for II in W?? ; do
+	    [ -d $II ] || continue
+	        TARFILE=my$II.tar.bz2
+		    TARFASC=$TARFILE.asc
+		        rm -f $TARFILE $TARFASC
+			    echo "tar cfj $TARFILE $II/"
+			        tar cfj $TARFILE $II/
+				    echo "gpg --armor --output $TARFASC --encrypt --recipient $REC1 --recipient $REC2 $TARFILE"
+				        gpg --armor --output $TARFASC --encrypt --recipient $REC1 --recipient $REC2 $TARFILE
+				done
+				popd
 
-echo "sha256sum $FILES > $SHA"
-sha256sum $FILES > $SHA
+				echo "mv -f $HOME/RESULT/myW*.tar.bz2.asc ."
+				mv -f $HOME/RESULT/myW*.tar.bz2.asc .
 
-echo "sha256sum -c $SHA"
-sha256sum -c $SHA
+				echo "rm -f $SHA $SHA.asc"
+				rm -f $SHA $SHA.asc
 
-echo "gpg -o $SHA.asc -a -sb $SHA"
-gpg -o $SHA.asc -a -sb $SHA
+				echo "sha256sum $FILES > $SHA"
+				sha256sum $FILES > $SHA
 
-echo "gpg --verify $SHA.asc $SHA"
-gpg --verify $SHA.asc $SHA
+				echo "sha256sum -c $SHA"
+				sha256sum -c $SHA
 
-git add .
-git commit -m "OS211 machffud"
-git push origin master
+				echo "gpg -o $SHA.asc -a -sb $SHA"
+				gpg -o $SHA.asc -a -sb $SHA
 
-exit 0
+				echo "gpg --verify $SHA.asc $SHA"
+				gpg --verify $SHA.asc $SHA
 
-# Mon Sep 28 21:05:04 WIB 2020
-# Tue 29 Sep 2020 11:02:39 AM WIB
+				git add .
+				git commit -m "OS211 machffud"
+				git push origin master
+
+
+				exit 0
+
